+++ app/controllers/api/v1/base_controller.rb
@@ -1,6 +1,9 @@
 class Api::V1::BaseController < ActionController::API
   include ActionController::HttpAuthentication::Token::ControllerMethods
+  include Rails.application.routes.url_helpers
+
   before_action :authenticate_user!
+  before_action :check_access_locked!
 
   private
 
@@ -12,5 +15,14 @@ class Api::V1::BaseController < ActionController::API
 
   def current_user
     @current_user
+  end
+
+  def check_access_locked!
+    return unless current_user&.access_locked?
+    render json: { error: "access_locked", payment_url: new_subscription_url }, status: :forbidden
+  end
+
+  def default_url_options
+    { host: Rails.application.credentials.dig(:app_host) }
   end
 end